refactor(manager): move ClusterInfo type to base
